% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{eckertIV}
\alias{eckertIV}
\title{Eckert IV Map Projection}
\usage{
eckertIV(dtLL, precision = 0.001, projColNames = c("x_EIV", "y_EIV"))
}
\arguments{
\item{dtLL}{A data.table with at minimum two numerical columns, one named 'lat' and the other 'lon', containing latitude and longitude points (in degrees) to be transformed.}

\item{precision}{The precision with which the projected values are calculated (the calculation is numerical using Newton's Method).}

\item{projColNames}{A length-2 character vector. In the returned data.table, what should the transformed columns be named?}
}
\description{
One of ReMIXTURE's three inbuilt map projection functions, the Eckert IV. See (https://en.wikipedia.org/wiki/Eckert_IV_projection)
}
