% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{equirectangular}
\alias{equirectangular}
\title{Equirectangular Map Projection}
\usage{
equirectangular(dtLL, projColNames = c("x_ER", "y_ER"))
}
\arguments{
\item{dtLL}{A data.table with at minimum two numerical columns, one named 'lat' and the other 'lon', containing latitude and longitude points (in degrees) to be transformed.}

\item{projColNames}{A length-2 character vector. In the returned data.table, what should the transformed columns be named?}
}
\description{
One of ReMIXTURE's three inbuilt map projection functions, the simple equirectangular projection. See (https://en.wikipedia.org/wiki/Equirectangular_projection)
}
