% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDescript.R
\docType{data}
\name{mapData10}
\alias{mapData10}
\title{Map data, 10 metre resolution}
\format{
A data.table giving the polygon outlines of map objects (countries, typically).
The columns are as follows:
lon (numeric) The longitude of a point
lat (numeric) The latitude of a point
sovereignty / name (character) Identifying information for each polygon from NaturalEarth. Not used by ReMIXTURE.
plotGrp (integer) Used to uniquely identify each polygon
ptIdx (integer) The order in which the points should be drawn/joined to draw the polygon correctly.
}
\source{
{ https://www.naturalearthdata.com/ }
}
\usage{
mapData10
}
\description{
This is used to draw the maps, and can be passed to \verb{$run()} via the \code{mapData} argument. In principle, a custom data.table with the same columns could be substituted to get fully customisable maps. The data come from NaturalEarth.
}
\keyword{datasets}
