% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReMIXTURE.R
\docType{package}
\name{ReMIXTURE-package}
\alias{ReMIXTURE}
\alias{ReMIXTURE-package}
\title{ReMIXTURE: Regionwise similarity analysis using a resampling/heirarchical clustering based method.}
\description{
Begins by using an algorithm to take data from collections of items with known inter-item distances and some 'region' property, calculate an overlap statistic between regions suitable to questions like "how similar is the collection of chilli peppers collected in this region to the collections found in other regions". Comparable values are given to describe the total "diversity" of items in each region, and the proportion of items that are unique to a region. Then it plots these overlaps, diversity, and uniqueness scores as lines connecting regions on a globe. An earlier version of this method is described in Rabanus-Wallace & Tripodi et al. (2021), this newer version differs significantly, with the algorithm updated to give much more intuitively interpretable results.
}
\section{Warning}{

This is under development, in early stages.
This version is significantly updated and differs non-trivially from the version presented in the paper Tripodi & Rabanus-Wallace et. al. 2021 about peppers. If you wish to use this method, please contact me to talk about how to cite it. A publication is coming soon.
}

\examples{

#installation
devtools::install_github("https://github.com/mtrw/ReMIXTURE")

See README on Github--documentation is coming soon.

}
\author{
\strong{Maintainer}: M Timothy Rabanus-Wallace \email{mtrw85@gmail.com}

}
\keyword{internal}
