% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadescript.R
\docType{data}
\name{ReMIXTURE_example_distance_matrix}
\alias{ReMIXTURE_example_distance_matrix}
\title{Matrix of distances between different samples}
\format{
A matrix with 972 rows and 972 columns. Matrix is diagonal.
Upper triangle and lower triangle are identical.
It is possible to populate one of the triangles with zeroes
}
\source{
{Rabanus-Wallace & Tripodi et al. (2021) }
}
\usage{
ReMIXTURE_example_distance_matrix
}
\description{
Dataset contains 972  chili hot pepper samples from each region
from Rabanus-Wallace & Tripodi et al. (2021) data.
This is a subset of the original data.
}
\keyword{datasets}
