% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadescript.R
\docType{data}
\name{ReMIXTURE_example_region_positions}
\alias{ReMIXTURE_example_region_positions}
\title{Metadata for map plotting}
\format{
A data table object with 11 rows and 4 variables:
\describe{
  \item{region}{A region, as stated in colnames or rownames in a distance matrix}
  \item{y}{y coordinates for a region center, i.e., standard latitude values in decimal format as you could read off Google maps}
  \item{x}{x coordinates for a region center, i.e., standard longitude values in decimal format as you could read off Google maps}
}
}
\source{
{Rabanus-Wallace & Tripodi et al. (2021) }
}
\usage{
ReMIXTURE_example_region_positions
}
\description{
A dataset contains metadata, used for map plotting.
}
\keyword{datasets}
