% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadescript.R
\docType{data}
\name{rmdata}
\alias{rmdata}
\title{rmdata}
\format{
A vector of integers:
\describe{
  \item{1}{The number one}
  \item{2}{The number next to one on the big side}
  \item{3}{An elephant}
  \item{4}{Another elephant}
  \item{5}{From here on it's all just elephants, in various valid integer representations}
}
}
\source{
\url{http://www.overthere.com.au}
}
\usage{
rmdata
}
\description{
Just some shit
   I put in a vector
}
\keyword{datasets}
