% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{run}
\alias{run}
\title{run}
\arguments{
\item{iterations}{[1000] The number of times subsamples are drawn, clustered, and counted.}

\item{subsample_proportions}{[0.8] size of subsample to draw each iteration. An equal number of samples will be selected from each region, the number being
\eqn{\text{subsample\_proportions} \times \text{\# of samples in region with the fewest samples}}
(rounded to the nearest integer).}

\item{h_cutoff}{["auto"] Controls the clustering cutoff \eqn{H}. There are three modes possible:
\itemize{
\item (Default) Random \eqn{H}, truncated normal distribution. Each iteration, \eqn{H} will be drawn from a normal distribution, truncated to the values given by \code{h_cutoff_normal_truncRange}. To use this option, you must provide the means of the distributions you want to try in \code{h_cutoff}, and the standard deviations (one for each mean) in \code{h_cutoff_normal_sigma}.
\itemize{
\item If \code{h_cutoff="auto"} (the default), then a single normal distribution will be used, with the same mean and sd as the distances in the distance matrix, equivalent to \code{remixture_obj$run(h_cutoff=mean(remixture_obj$distance_matrix),h_cutoff_normal_sigma=sd(remixture_obj$distance_matrix))}
}
\item Fixed \eqn{H}. To use this option, provide a list of values to try as \eqn{H}. A run will be conducted with each, using the same value at each iteration. If you use this mode, you should clearly indicate it in any figure or publication, e.g. "ReMIXTURE plot using truncated random normal \eqn{H} (mean = ..., sd = ..., range=[...,...])".
\item Random \eqn{H}, empirical distribution. This samples values of \eqn{H} from the distances in the distance matrix. This should probably never be used, because while it does help produce meaningful plots in cases of extremely structured populations, the results are just not very intuitive. If you use this mode, you should clearly indicate it in any figure or publication, e.g. "ReMIXTURE plot using empirically distributed \eqn{H}".
}}

\item{h_cutoff_normal_sigma}{[NULL] See description for \code{h_cutoff}.}

\item{h_cutoff_normal_truncRange}{[range(self$distance_matrix)] See description for \code{h_cutoff}.}

\item{diagnosticPlotMDSclusters}{[FALSE] A diagnostic tool. If true, at each iteration, will draw the subsampled samples on an MDS plot, and draw hulls around the clusters.}
}
\description{
Run the ReMIXTURE algorithm and save the results in the object.
}
\examples{
(1+1)==2
}
