% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReMIXTURE.R
\name{sayname}
\alias{sayname}
\title{sayname}
\usage{
sayname(x = "nothing")
}
\arguments{
\item{x}{Anything printable.}
}
\value{
Nothing. Just prints nothing interesting.
}
\description{
Says the name of the intended future of this package
}
\section{Warning}{

Does nothing interesting.
}

\examples{
sayname("Humboldt")
}
