% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{upper_tri_ply}
\alias{upper_tri_ply}
\title{Apply a function FUN to a vector made from the entries in the upper triangle of the matrix of m.}
\usage{
upper_tri_ply(m, FUN, ...)
}
\arguments{
\item{m}{[no default] A matrix.}

\item{FUN}{[no default] The function to apply.}

\item{...}{Extra arguments passed to FUN( ... )}
}
\value{
The output as described
}
\description{
Apply a function FUN to a vector made from the entries in the upper triangle of the matrix of m.
}
