% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{winkelIII}
\alias{winkelIII}
\title{Winkel III map projection}
\usage{
winkelIII(dtLL, projColNames = c("x_W3", "y_W3"))
}
\arguments{
\item{dtLL}{A data.table with at minimum two numerical columns, one named 'lat' and the other 'lon', containing latitude and longitude points (in degrees) to be transformed.}

\item{projColNames}{A length-2 character vector. In the returned data.table, what should the transformed columns be named?}
}
\description{
One of ReMIXTURE's three inbuilt map projection functions, the Winkel Tripel aka Winkel III. See (https://en.wikipedia.org/wiki/Winkel_tripel_projection)
}
