% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReMIXTURE.R, R/class.R, R/getset.R
\docType{_PACKAGE}
\name{ReMIXTURE}
\alias{ReMIXTURE}
\title{ReMIXTURE: Regionwise similarity analysis using a resampling/heirarchical clustering based method.}
\format{
An active binding in an R6 class.
}
\usage{
ReMIXTURE_object$region_table
}
\description{
ReMIXTURE: Regionwise similarity analysis using a resampling/heirarchical clustering based method.
}
\section{Fields}{

\describe{
\item{\code{region_table}}{Get the region table from a ReMIXTURE object}
}}

\section{Warning}{

This is under development, in early stages.
This version is significantly updated and differs non-trivially from the version presented in the paper Tripodi & Rabanus-Wallace et. al. 2021 about peppers. If you wish to use this method, please contact me to talk about how to cite it. A publication is coming soon.
}

\examples{

#installation
devtools::install_github("https://github.com/mtrw/ReMIXTURE")

See README on Github--documentation is coming soon.

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NA-run}{\code{ReMIXTURE$run()}}
\item \href{#method-NA-new}{\code{ReMIXTURE$new()}}
\item \href{#method-NA-plot_heatmaps}{\code{ReMIXTURE$plot_heatmaps()}}
\item \href{#method-NA-plot_clustercounts}{\code{ReMIXTURE$plot_clustercounts()}}
\item \href{#method-NA-plot_h_optimisation}{\code{ReMIXTURE$plot_h_optimisation()}}
\item \href{#method-NA-plot_maps}{\code{ReMIXTURE$plot_maps()}}
\item \href{#method-NA-plot_distance_densities}{\code{ReMIXTURE$plot_distance_densities()}}
\item \href{#method-NA-plot_MDS}{\code{ReMIXTURE$plot_MDS()}}
\item \href{#method-unknown-clone}{\code{ReMIXTURE$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-run"></a>}}
\if{latex}{\out{\hypertarget{method-NA-run}{}}}
\subsection{Method \code{run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReMIXTURE$run(
  iterations = 100,
  subsample_proportions = c(0.8),
  h_cutoffs = seq(min(self$dm), max(self$dm), length.out = 10),
  diagnosticPlotMDSclusters = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-new"></a>}}
\if{latex}{\out{\hypertarget{method-NA-new}{}}}
\subsection{Method \code{new()}}{
Create a new ReMIXTURE object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReMIXTURE$new(distance_matrix, region_table)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{distance_matrix}}{An all-vs-all, full numeric distance matrix, with rownames and colnames giving the region of origin of the corresponding individual.}

\item{\code{region_table}}{A data.table describing the longitudes/latitudes of each region, with columns named "region" (character), and "lon" and "lat" (numeric or integer). The "region" column must have names corresponding to all the row/column names of the distance matrix.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a new \code{ReMIXTURE} object.
penis = function(){}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-plot_heatmaps"></a>}}
\if{latex}{\out{\hypertarget{method-NA-plot_heatmaps}{}}}
\subsection{Method \code{plot_heatmaps()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReMIXTURE$plot_heatmaps(colPalette = c("#f2f5ff", "#000570"), ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-plot_clustercounts"></a>}}
\if{latex}{\out{\hypertarget{method-NA-plot_clustercounts}{}}}
\subsection{Method \code{plot_clustercounts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReMIXTURE$plot_clustercounts(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-plot_h_optimisation"></a>}}
\if{latex}{\out{\hypertarget{method-NA-plot_h_optimisation}{}}}
\subsection{Method \code{plot_h_optimisation()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReMIXTURE$plot_h_optimisation(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-plot_maps"></a>}}
\if{latex}{\out{\hypertarget{method-NA-plot_maps}{}}}
\subsection{Method \code{plot_maps()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReMIXTURE$plot_maps(
  run = NULL,
  focalRegion = NULL,
  range_lon = c(-179, 179),
  range_lat = c(-85, 85),
  width_lims = c(5, 13),
  alpha_lims = c(0.05, 0.99),
  projection = eckertIV,
  curvature_matrix = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-plot_distance_densities"></a>}}
\if{latex}{\out{\hypertarget{method-NA-plot_distance_densities}{}}}
\subsection{Method \code{plot_distance_densities()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReMIXTURE$plot_distance_densities(
  set_bw = 0.001,
  set_xlims = c(0, 1),
  samePlot = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-plot_MDS"></a>}}
\if{latex}{\out{\hypertarget{method-NA-plot_MDS}{}}}
\subsection{Method \code{plot_MDS()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReMIXTURE$plot_MDS(
  PCs = c(1L, 2L),
  colPalette = c("#88000088", "#88880088", "#00880088", "#00008888", "#88008888"),
  distanceMatrix = NULL,
  showLegend = TRUE,
  doPlot = TRUE,
  pch = 20,
  cex = 0.4,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-unknown-clone"></a>}}
\if{latex}{\out{\hypertarget{method-unknown-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReMIXTURE$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
