% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReMIXTURE.R, R/class.R
\docType{package}
\name{ReMIXTURE}
\alias{ReMIXTURE}
\title{ReMIXTURE: Regionwise similarity analysis using a resampling/heirarchical clustering based method.}
\section{Warning}{

This is under development, in early stages.
This version is significantly updated and differs non-trivially from the version presented in the paper Tripodi & Rabanus-Wallace et. al. 2021 about peppers. If you wish to use this method, please contact me to talk about how to cite it. A publication is coming soon.
}

\examples{

#installation
devtools::install_github("https://github.com/mtrw/ReMIXTURE")

#new analysis using example data provided
my_analysis <- ReMIXTURE$new(
  distance_matrix = ReMIXTURE::ReMIXTURE_example_distance_matrix,
  region_table = ReMIXTURE::ReMIXTURE_example_region_positions
)

#A good selection of h-cutoffs should span the lower end of all distance peaks
my_analysis$plot_distance_densities(set_xlims = c(0,0.2))

#Run the analysis
my_analysis$run(iterations = 500,subsample_proportions = c(0.7,0.8,0.9),h_cutoffs=seq(.008,.02,l=6))

#Raw results
my_analysis$run_results

#Tools to assess h-cutoff values
   #Good values are often found between where the median clustercounts are equal, and where the inter-region cluster counts maximise
my_analysis$plot_clustercount_diag_nondiag_means()
   #A good value has a "balanced" heatmap, the main features in which are stable at nearby values
my_analysis$plot_heatmaps()
   #A good value does not incur many or any per-region cluster counts near one. A few may be inevitable in some datasets though.
my_analysis$plot_clustercounts()

#Plot the regions on the globe, and add raw plotting data to my_analysis$run_results
my_analysis$plot_maps(run = 9,alpha_norm_per_region = T)

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-run}{\code{ReMIXTURE$run()}}
\item \href{#method-new}{\code{ReMIXTURE$new()}}
\item \href{#method-plot_heatmaps}{\code{ReMIXTURE$plot_heatmaps()}}
\item \href{#method-plot_clustercounts}{\code{ReMIXTURE$plot_clustercounts()}}
\item \href{#method-plot_h_optimisation}{\code{ReMIXTURE$plot_h_optimisation()}}
\item \href{#method-plot_maps}{\code{ReMIXTURE$plot_maps()}}
\item \href{#method-plot_distance_densities}{\code{ReMIXTURE$plot_distance_densities()}}
\item \href{#method-clone}{\code{ReMIXTURE$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run"></a>}}
\if{latex}{\out{\hypertarget{method-run}{}}}
\subsection{Method \code{run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReMIXTURE$run(
  iterations = 2500,
  subsample_proportions = c(0.8, 0.9),
  h_cutoffs = seq(0.001, 0.2, l = 6),
  colpalette_heatmap = colorRampPalette(c("#f2f5ff", "#214feb", "#001261"))(100),
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new ReMIXTURE object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReMIXTURE$new(distance_matrix, region_table)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{distance_matrix}}{An all-vs-all, full numeric distance matrix, with rownames and
colnames giving the region of origin of the corresponding individual.}

\item{\code{region_table}}{A data.table describing the lat(y)/long(x)s of each region (numeric or integer values), with columns named "region", "x" and "y".}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a new \code{ReMIXTURE} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_heatmaps"></a>}}
\if{latex}{\out{\hypertarget{method-plot_heatmaps}{}}}
\subsection{Method \code{plot_heatmaps()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReMIXTURE$plot_heatmaps(
  colpalette = colorRampPalette(c("#f2f5ff", "#214feb", "#001261"))(100),
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_clustercounts"></a>}}
\if{latex}{\out{\hypertarget{method-plot_clustercounts}{}}}
\subsection{Method \code{plot_clustercounts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReMIXTURE$plot_clustercounts(
  colpalette = colorRampPalette(c("#f2f5ff", "#214feb", "#001261"))(100),
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_h_optimisation"></a>}}
\if{latex}{\out{\hypertarget{method-plot_h_optimisation}{}}}
\subsection{Method \code{plot_h_optimisation()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReMIXTURE$plot_h_optimisation(
  colpalette = colorRampPalette(c("#f2f5ff", "#214feb", "#001261"))(100),
  plot_entropy = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_maps"></a>}}
\if{latex}{\out{\hypertarget{method-plot_maps}{}}}
\subsection{Method \code{plot_maps()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReMIXTURE$plot_maps(
  run = NULL,
  alpha_norm_per_region = NULL,
  alpha_correlation = F,
  lon_angle = 0,
  lat_angle = 0,
  width_lims = c(5, 35),
  alpha_lims = c(0.05, 0.99)
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_distance_densities"></a>}}
\if{latex}{\out{\hypertarget{method-plot_distance_densities}{}}}
\subsection{Method \code{plot_distance_densities()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReMIXTURE$plot_distance_densities(set_bw = 0.001, set_xlims = c(0, 1))}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReMIXTURE$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
