% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{ReMIXTURE}
\alias{ReMIXTURE}
\title{ReMIXTURE}
\description{
ReMIXTURE is a tool for visualising genetic (or any) diversity across space.

It produces intuitive (and pretty!) plots that show the diversity in different geographic regions, and how it is shared between regions.
These plots are based on ReMIXTURE's diversity metric, which is designed to be maximally intuitive. Basically the plots are designed work like Venn diagrams: Diversity in a group is partitioned into diversity that is unique to that group, and that which overlaps other groups--and this is quantified for every focal--target group pair.

ReMIXTURE is designed for data where many samples exist from many (or ... at least a few) locations. It requires just a distance matrix describing the difference between every sample pair, and a table describing where on Earth each sample is located (latitude and longitude).

This is the first release version of the package, and development is ongoing.

For a short lesson on how to run ReMIXTURE, follow the tutorial in the examples section below.
}
\examples{
#### A SHORT TUTORIAL ####
library(ReMIXTURE)

# Create ReMIXTURE analysis object from an included dataset (Tripodi & Rabanus-Wallace et al. 2021. PNAS).
rm <- ReMIXTURE$new(
  distance_matrix = ReMIXTURE_example_distance_matrix,
  region_table = ReMIXTURE_example_region_table
)

# An MDS plot is a quick way to get a feel for the structure of the data.
rm$plot_MDS()
#
# A run involves taking a subsample of all the samples in the distance matrix, such that every region has the same number of samples included. This subsample is hierarchically clustered (imagine creating a dendrogram and cutting it at a certain height). Clusters are then counted. The number of clusters in which a region appears is a proxy for that region's total genetic diversity. The number of clusters in which *only* one region appears is a proxy for the diversity unique to that region. The number of clusters in which members of a pair of clusters appears is a proxy for the diversity that is overlapping between those two regions. the concept is similar to the way a Venn diagram might work. This is done `iterations` times, and all these various counts are averaged at the end.
#
# How is it decided how high the tree is cut? The details are in `?ReMIXTURE` under '$run'.
# Ideally, a collection of parameter settings are chosen and the results assessed to choose a good one. By default, the run function tries to choose a good selection of values, and it does a good job.
rm$run(iterations = 25) # Iterations set low as we will be doing several runs and comparing results, so we can sacrifice a little accuracy for now.

# For in an intuitive look at how the diversity metric changes with the parameter settings, view them simply as a grid, for each run (press <enter> to get through them). Notice how with low H values, unique diversity is over-emphasised and overlapping diversity is under-emphasised
rm$plot_results_grid()

# A good balance can be found by summing region-unique/-overlapping clusters. At some point they are expected to cross over.
rm$plot_h_optimisation()

# Note also that the number of multi-region clusters reaches a maximum, when the H is big enough to produce many such clusters, but before the clustersbecome so large that the total cluster number decline enough to drag this number down.
# H values between the crossover point and the multi-region maximum are usually good for showing a balanced representation. I prefer the crossover point, here about run 4.
# Some more diagnosis can be done, but for now let's produce our remixture maps!
# BEWARE: What we are doing is great for a demo but before you publish you should be sure to do the rest of this tutorial.
rm$plot_maps(run = 4, focalRegion = "Africa")
  # Let's make those diversity visualisation bigger and get some curved lines ...
set.seed(42.42)
rm$plot_maps(
  run=4,
  focalRegion = "Africa",
  width_max = 30,
  curvature_matrix = "random"
)
  # Let's crop the map and try another projection ...
rm$plot_maps(
  run=4,
  focalRegion = "Africa",
  width_max = 30,
  curvature_matrix = "random",
  range_lon = c(-120,140),
  range_lat = c(-30,60),
  projection = equirectangular
)
  # Finally let's do this as a plot with all maps using the standard R layout options. I will also show each region's diversity metrics only when it is the focal region, using `diversityCirclesFocalOnly = TRUE`.
par(mfrow=c(4,4),mar=c(0.2,0.2,1,0.2))
rm$plot_maps(
  run=4,
  #focalRegion = "Africa", # without this argument, will sequentially plot maps for all focal regions
  width_max = 30,
  curvature_matrix = "random",
  range_lon = c(-120,140),
  range_lat = c(-30,60),
  projection = equirectangular,
  diversityCirclesFocalOnly = TRUE
)
# clear plots and reset parameters
dev.off()

#### More functions and details ####

# Extract info from the object
rm$distance_matrix[1:10,1:10] # get back the distance matrix
rm$region_table # get back the region table
rm$mds # get back the mds data

# Diagnostic tools -- the distribution of distances between/within different regions, and how they relate to the distribution of H
rm$plot_distance_densities(HdistFromRun=4) # will plot all sequentially including H distribution as used in run 4
rm$plot_distance_densities( # overlay a hypothetical H distribution of your choosing
  H = 0.01,
  H_truncNorm_sd = 0.02,
  H_truncNorm_lims = c(.005,.02)
)
rm$plot_distance_densities( # as above, but with a constant H
  H = 0.01
)
# We can plot MDS plots as we perform a run, and draw hulls surrounding the clusters. This is great for giving an impression of how the subsampling level looks and whether there are any peculiarities interfering with the clustering steps.
# Note for this run we have chosen a constant H, as shown in the last plot
# Once you have seen enough iterations, hit 'f' to finish the run without plotting.
rm$run(
  iterations = 100,
  H = 0.01,
  diagnosticPlotMDSclusters = TRUE
)

# That's it!


## ------------------------------------------------
## Method `ReMIXTURE$new`
## ------------------------------------------------

rm <- ReMIXTURE$new(
  distance_matrix = ReMIXTURE_example_distance_matrix,
  region_table = ReMIXTURE_example_region_table
)
rm$distance_matrix[1:10,1:10]
rm$region_table
rm <- ReMIXTURE$new(
  distance_matrix = ReMIXTURE_example_distance_matrix,
  region_table = ReMIXTURE_example_region_table
)
rm$region_table
rm$distance_matrix[1:10,1:10]
# Using build-in ReMIXTURE test datasets
rm <- ReMIXTURE$new(
  distance_matrix = ReMIXTURE_example_distance_matrix,
  region_table = ReMIXTURE_example_region_table
)
rm$run() # Default run settings
rm$run_results
# Using build-in ReMIXTURE test datasets
rm <- ReMIXTURE$new(
  distance_matrix = ReMIXTURE_example_distance_matrix,
  region_table = ReMIXTURE_example_region_table
)
rm$plot_MDS() # Default settings
rm$mds
# Using build-in ReMIXTURE test datasets
rm <- ReMIXTURE$new(
  distance_matrix = ReMIXTURE_example_distance_matrix,
  region_table = ReMIXTURE_example_region_table
)
rm$distance_matrix[1:10,1:10]
rm$region_table

## ------------------------------------------------
## Method `ReMIXTURE$run`
## ------------------------------------------------

# See examples section of ?ReMIXTURE

## ------------------------------------------------
## Method `ReMIXTURE$plot_results_grid`
## ------------------------------------------------

# See ?ReMIXTURE

## ------------------------------------------------
## Method `ReMIXTURE$plot_clustercounts`
## ------------------------------------------------

# See ?ReMIXTURE

## ------------------------------------------------
## Method `ReMIXTURE$plot_h_optimisation`
## ------------------------------------------------

# See ?ReMIXTURE

## ------------------------------------------------
## Method `ReMIXTURE$plot_maps`
## ------------------------------------------------

# See ?ReMIXTURE

## ------------------------------------------------
## Method `ReMIXTURE$plot_distance_densities`
## ------------------------------------------------

# See ?ReMIXTURE

## ------------------------------------------------
## Method `ReMIXTURE$plot_MDS`
## ------------------------------------------------

# See ?ReMixture
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{distance_matrix}}{Return the distance matrix.}

\item{\code{region_table}}{Return the region table with lat/lon information.}

\item{\code{run_results}}{Return the raw results of ReMIXTURE runs (runs must be first performed with \verb{<ReMIXTURE_object>$run()})}

\item{\code{mds}}{Return the mds data (created when \verb{<ReMIXTURE_object>$plotMDS()} is run)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ReMIXTURE-new}{\code{ReMIXTURE$new()}}
\item \href{#method-ReMIXTURE-run}{\code{ReMIXTURE$run()}}
\item \href{#method-ReMIXTURE-plot_results_grid}{\code{ReMIXTURE$plot_results_grid()}}
\item \href{#method-ReMIXTURE-plot_clustercounts}{\code{ReMIXTURE$plot_clustercounts()}}
\item \href{#method-ReMIXTURE-plot_h_optimisation}{\code{ReMIXTURE$plot_h_optimisation()}}
\item \href{#method-ReMIXTURE-plot_maps}{\code{ReMIXTURE$plot_maps()}}
\item \href{#method-ReMIXTURE-plot_distance_densities}{\code{ReMIXTURE$plot_distance_densities()}}
\item \href{#method-ReMIXTURE-plot_MDS}{\code{ReMIXTURE$plot_MDS()}}
\item \href{#method-ReMIXTURE-clone}{\code{ReMIXTURE$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReMIXTURE-new"></a>}}
\if{latex}{\out{\hypertarget{method-ReMIXTURE-new}{}}}
\subsection{Method \code{new()}}{
Create a new ReMIXTURE object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReMIXTURE$new(distance_matrix, region_table)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{distance_matrix}}{[no default] An all-vs-all, full numeric distance matrix, with rownames and colnames giving the region of origin of the corresponding individual.}

\item{\code{region_table}}{[no default] A data.table describing the longitudes/latitudes of each region, with columns named "region" (character), and "lon" and "lat" (numeric or integer). The "region" column must have names corresponding to all the row/column names of the distance matrix.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new ReMIXTURE object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{rm <- ReMIXTURE$new(
  distance_matrix = ReMIXTURE_example_distance_matrix,
  region_table = ReMIXTURE_example_region_table
)
rm$distance_matrix[1:10,1:10]
rm$region_table
}
\if{html}{\out{</div>}}

\if{html}{\out{<div class="r example copy">}}
\preformatted{rm <- ReMIXTURE$new(
  distance_matrix = ReMIXTURE_example_distance_matrix,
  region_table = ReMIXTURE_example_region_table
)
rm$region_table
rm$distance_matrix[1:10,1:10]
}
\if{html}{\out{</div>}}

\if{html}{\out{<div class="r example copy">}}
\preformatted{# Using build-in ReMIXTURE test datasets
rm <- ReMIXTURE$new(
  distance_matrix = ReMIXTURE_example_distance_matrix,
  region_table = ReMIXTURE_example_region_table
)
rm$run() # Default run settings
rm$run_results
}
\if{html}{\out{</div>}}

\if{html}{\out{<div class="r example copy">}}
\preformatted{# Using build-in ReMIXTURE test datasets
rm <- ReMIXTURE$new(
  distance_matrix = ReMIXTURE_example_distance_matrix,
  region_table = ReMIXTURE_example_region_table
)
rm$plot_MDS() # Default settings
rm$mds
}
\if{html}{\out{</div>}}

\if{html}{\out{<div class="r example copy">}}
\preformatted{# Using build-in ReMIXTURE test datasets
rm <- ReMIXTURE$new(
  distance_matrix = ReMIXTURE_example_distance_matrix,
  region_table = ReMIXTURE_example_region_table
)
rm$distance_matrix[1:10,1:10]
rm$region_table
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReMIXTURE-run"></a>}}
\if{latex}{\out{\hypertarget{method-ReMIXTURE-run}{}}}
\subsection{Method \code{run()}}{
Run the ReMIXTURE algorithm and save the results in the object. Multiple runs can be requested to aid parameter selection (see description for \code{H}), which is strongly recommended.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReMIXTURE$run(
  iterations = 1000,
  subsample_proportions = c(0.8),
  H = "auto",
  H_truncNorm_sd = NULL,
  H_truncNorm_range = upper_tri_ply(private$m, range),
  diagnosticPlotMDSclusters = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{iterations}}{[1000] The number of times subsamples are drawn, clustered, and counted.}

\item{\code{subsample_proportions}}{[0.8] Size of subsample to draw each iteration. An equal number of samples will be selected from each region, the number being
\eqn{\text{subsample\_proportions} \times \text{\# of samples in region with the fewest samples}}
(rounded to the nearest integer).}

\item{\code{H}}{["auto"] Controls the clustering cut heights \eqn{H} used in each run. There are three modes possible:
\itemize{
\item (Default) Random \eqn{H}, truncated normal distributions. Each iteration, \eqn{H} will be drawn from a truncated normal distribution, truncated to the values given by \code{H_truncNorm_range}. To use this option, you must provide the means of the distributions you want to try in \code{H}, and the standard deviations (one corresponding to each mean) in \code{H_truncNorm_sd}.
\itemize{
\item If \code{H="auto"} (the default), then a selection of 16 normal distributions will be chosen based on a simple heuristic that typically captures some very good values. Briefly, the values of \code{H} fully span the interguartile ranges of inter-sample distances of each region, \code{H_truncNorm_sd} is half the standard deviation of inter-sample distances in the region with the smallest such value (i.e. it is given the same value in all 16 distributions), and \code{H_truncNorm_range} is just the range of all inter-sample distances in the whole dataset.
}
\item Fixed \eqn{H}. To use this option, provide a list of values to try as \eqn{H}. A run will be conducted with each, using the same value at each iteration.
\item Random \eqn{H}, empirical distribution. This samples values of \eqn{H} from the distances in the distance matrix. This should probably never be used, because while it does help produce meaningful plots in cases of extremely structured populations, the results are just not very intuitive. If you use this mode, you should clearly indicate it in any figure or publication, e.g. "ReMIXTURE plot using empirically distributed \eqn{H}". But just don't. This option will be removed.
}}

\item{\code{H_truncNorm_sd}}{[NULL] See description for \code{H}.}

\item{\code{H_truncNorm_range}}{[upper_tri_ply(private$m,range)] See description for \code{H}. By default it is set to the range of values in the distance matrix.}

\item{\code{diagnosticPlotMDSclusters}}{[FALSE] A very useful diagnostic tool, especially for weirdly-structured datasets. If true, at each iteration, will draw the subsampled samples on an MDS plot, and draw hulls around the clusters.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NULL
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# See examples section of ?ReMIXTURE
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReMIXTURE-plot_results_grid"></a>}}
\if{latex}{\out{\hypertarget{method-ReMIXTURE-plot_results_grid}{}}}
\subsection{Method \code{plot_results_grid()}}{
Plot the unique and overlapped diversity recorded in a ReMIXTURE run, in something like a heatmap format but using the circle-area conventions as per \code{plot_maps()}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReMIXTURE$plot_results_grid(runs = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{runs}}{[NULL] For which runs would you like a plot? An integer vector. By default, all runs.}

\item{\code{...}}{Additional arguments ultimately passed to \code{base::plot()}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Nothing
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# See ?ReMIXTURE
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReMIXTURE-plot_clustercounts"></a>}}
\if{latex}{\out{\hypertarget{method-ReMIXTURE-plot_clustercounts}{}}}
\subsection{Method \code{plot_clustercounts()}}{
Plot the cluster counts for each region in each run. This is useful for assessing how different values of \eqn{H} affect the output, in particular checking there aren't regions with very low cluster counts, or that they are not "maxxed out", either of which will probably give dodgy results unreliable.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReMIXTURE$plot_clustercounts()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Nothing
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# See ?ReMIXTURE
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReMIXTURE-plot_h_optimisation"></a>}}
\if{latex}{\out{\hypertarget{method-ReMIXTURE-plot_h_optimisation}{}}}
\subsection{Method \code{plot_h_optimisation()}}{
These counts are excellent for determining a good value for \eqn{H}, in cases where the default choice is not desirable. A good value typically gives a nice balance of multi- and single-region clusters, typically between where the two values cross over and (as occurs in most datasets) a point where the number of multi-region clusters hits some maximum.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReMIXTURE$plot_h_optimisation()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Nothing
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# See ?ReMIXTURE
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReMIXTURE-plot_maps"></a>}}
\if{latex}{\out{\hypertarget{method-ReMIXTURE-plot_maps}{}}}
\subsection{Method \code{plot_maps()}}{
Produces the canonical ReMIXTURE plots.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReMIXTURE$plot_maps(
  run = NULL,
  focalRegion = NULL,
  range_lon = c(-179, 179),
  range_lat = c(-85, 85),
  width_max = 10,
  alpha_max = 1,
  diversityCirclesFocalOnly = FALSE,
  projection = eckertIV,
  curvature_matrix = NULL,
  mapData = mapData110
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{run}}{[NULL] If multiple runs were done, choose which run to use. This is best assessed using \verb{<ReMIXTURE Object>$plot_h_optimisation()}, \verb{<ReMIXTURE Object>$plot_results_grid()}, and \verb{<ReMIXTURE Object>$plot_clustercounts()}.}

\item{\code{focalRegion}}{[NULL] A character string naming one of the regions as it occurs in the region table, chosen as the focal region (from which lines showing inter-region overlap will emanate). If left NULL then a map will be plotted with each region as the focus in turn--this is best used after setting up a multi-panel plot with some variant of \verb{par(mfrow=c(<number of rows>,<number of columns>))}.}

\item{\code{range_lon}}{[c(-179.0,179.0)] Limit the map to some range of longitudes. Must be a vector of 2 numbers. For awkward reasons, using 180 or -180 can cause graphical bugs and I am deeply sorry.}

\item{\code{range_lat}}{[c(-85.0,85.0)] Limit the map to some range of latitudes. Must be a vector of 2 numbers.}

\item{\code{width_max}}{[15.0] The maximum width of the circles/lines, in units of lat/lons. This width will correspond to the highest cluster count and the others will be scaled such that zero clusters <=> zero width.}

\item{\code{alpha_max}}{[1.0] As per width_max but controlling the alpha of connecting lines. Setting to NULL will disable alpha and make all lines solid. Can be set above 1.0, with weird results--probably don't do this.}

\item{\code{diversityCirclesFocalOnly}}{[FALSE] Plot the circle representing a region's total/unique/shared diversity at only the focal region. Otherwise, all plots will show the diversities at all regions. This is good when you have multiple plots on the page, see description for \code{focalRegion}.}

\item{\code{projection}}{[EckertIV] A function the performs projection of the map coordinates. Included in the package are \code{eckertIV}, \code{winkelIII}, and \code{equirectangular}. A custom function can be given--see details.}

\item{\code{curvature_matrix}}{[NULL] A matrix describing how the lines emanating from the focal region should bend. The (i,j)th entry gives the bend angle (in radians) of the line beginning at region i and ending at region j (indexed as per the order in the region table). In practice this is pretty tedious to enter manually. The argument "random" will auto-generate a matrix filled with entries ~ normal(0,0.3), which usually works well.}

\item{\code{mapData}}{[mapData110] The polygon data from which the map is made. For higher resolution, use mapData10. See e.g. ?mapData110.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
A custom projection function should follow these rules: its first argument, named 'dtLL' should take a data.table with columns 'lon' and 'lat'. The second argument, named 'projColNames', should take a length-2 character vector. The output should be a data.table with columns containing the transformed longitude and latitude values. These columns should be named as per the entries in the input 'projColNames'.
}

\subsection{Returns}{
Nothing
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# See ?ReMIXTURE
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReMIXTURE-plot_distance_densities"></a>}}
\if{latex}{\out{\hypertarget{method-ReMIXTURE-plot_distance_densities}{}}}
\subsection{Method \code{plot_distance_densities()}}{
Plot inter-sample distances as a density plot over the distance matrix--basically, a convenient way to judge how far apart samples tend to be, and thus how the clustering might behave when various H values (\eqn{H}) are used.

Each plot tells the story from the perspective of a focal region. Each region is assigned a colour that remains constant across plots. A region's samples' distances to each other are marked with a thick line allowing you to see what colour is for what region.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReMIXTURE$plot_distance_densities(
  set_bw = 0.001,
  set_xlims = upper_tri_ply(private$m, range),
  samePlot = FALSE,
  H = NULL,
  H_truncNorm_sd = NULL,
  H_truncNorm_lims = NULL,
  HdistFromRun = NULL,
  plotLegend = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{set_bw}}{[NULL] The bandwidth parameter 'bw' passed to \code{density()}. By default it will be set automatically.}

\item{\code{set_xlims}}{[NULL] Limits on the x-axis. Default is set by \code{density()}.}

\item{\code{samePlot}}{[FALSE] If TRUE, it will put all the plots on one page.}

\item{\code{H}}{[NULL] Visualise chosen values of (or distribution over) \eqn{H} on the plot. If only this argument is used, a single line will indicate that H value.}

\item{\code{H_truncNorm_sd}}{[NULL] Visualise a given distribution of \eqn{H} on the plot. If this argument is used (in conjunction with \code{H}), a truncated normal distribution will be shown.}

\item{\code{H_truncNorm_lims}}{[upper_tri_ply(private$m,range)] Visualise your values of \eqn{H} on the plot. If this argument is used, it will set the truncation limits of the normal distribution. By default it will be the range of inter-sample distances in the distance matrix.}

\item{\code{HdistFromRun}}{[NULL] Show the H distribution for a particular run. Obviously, this requires that a run has been done.}

\item{\code{plotLegend}}{[FALSE] If true, the last plot produced will be the colour legend.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisibly returns a data.table giving the legend colours.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# See ?ReMIXTURE
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReMIXTURE-plot_MDS"></a>}}
\if{latex}{\out{\hypertarget{method-ReMIXTURE-plot_MDS}{}}}
\subsection{Method \code{plot_MDS()}}{
An MDS plot showing relationships among samples

An MDS ordination plot (in the same flavour as a PCA--in fact in some circumstances they will be the same depending on what your distances represent), useful as a first-pass assessment of the relationships among your samples--what overlaps what, what is over/underrepresented in the sample set, etc.

These can take a very long time if the number of samples is high. In these cases, I suggest taking a smaller sample of the whole dataset.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReMIXTURE$plot_MDS(
  axes = c(1L, 2L),
  colPalette = c("#DD000088", "#DDDD0088", "#00DD0088", "#0000DD88", "#DD00DD88"),
  showLegend = TRUE,
  doPlot = TRUE,
  pch = 20,
  cex = 0.8,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{axes}}{[c(1,2)] A vector of the MDS axes that should be outputted.}

\item{\code{colPalette}}{[c("#DD000088","#DDDD0088","#00DD0088","#0000DD88","#DD00DD88")] A vector collection of colours that will be interpolated and applied to the different}

\item{\code{showLegend}}{[TRUE] Add a legend?}

\item{\code{doPlot}}{[TRUE] Produce a plot? If not, will still return the data for the MDS.}

\item{\code{pch}}{[20] The shape of the points. See \code{?points}.}

\item{\code{cex}}{[0.8] The size of the points. See \code{?points}.}

\item{\code{...}}{Other arguments passed to \code{plot()}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisibly returns a 2-list containing the (1) results of the MDS in a data.table and (2) information about the legend (another data.table).
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# See ?ReMixture
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReMIXTURE-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ReMIXTURE-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReMIXTURE$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
