% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReMIXTURE.R
\name{ReMIXTURE}
\alias{ReMIXTURE}
\title{Regionwise similarity analysis using a resampled nearest-neighbour method.}
\value{
A ReMIXTURE class object.
}
\description{
Regionwise similarity analysis using a resampled nearest-neighbour method.

Regionwise similarity analysis using a resampled nearest-neighbour method.
}
\section{Warning}{

This R6 implementation is under development. Occasional bugs are still coming to light.
}

\examples{
my_analysis <- ReMIXTURE$new(
distance_matrix = ReMIXTURE::distance_matrix,
   info_table = ReMIXTURE::geocenters
)

#Takes about 5--10 minutes without increasing the parallelise argument (avilable on linux-based systems).
my_analysis$run(iterations = 500,parallelize = 1)

my_analysis$plot_heatmap()

m <- my_analysis$plot_maps()
dev.off()
pdf("MY_PLOTS.pdf",width=5,height=50)
print(m)
dev.off()
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{distance_matrix}}{Return the distance matrix as a data.table}

\item{\code{results}}{Return the numerical results of $run() as a data.table}

\item{\code{resample_matrix}}{Return the results of resampling conducted by $run_resampling() as a data.table}

\item{\code{info_table}}{Return the information table that assigns latitudes, longitudes, and colours to the regions as a data.table}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ReMIXTURE$new()}}
\item \href{#method-run}{\code{ReMIXTURE$run()}}
\item \href{#method-run_resampling}{\code{ReMIXTURE$run_resampling()}}
\item \href{#method-plot_heatmap}{\code{ReMIXTURE$plot_heatmap()}}
\item \href{#method-plot_maps}{\code{ReMIXTURE$plot_maps()}}
\item \href{#method-clone}{\code{ReMIXTURE$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new ReMIXTURE object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReMIXTURE$new(distance_matrix, info_table = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{distance_matrix}}{An all-vs-all, full numeric distance matrix, with rownames and
colnames giving the region of origin of the corresponding individual.}

\item{\code{info_table}}{A data.table rescribing the lat(y)/long(x)s of each region, with columns named "region", "x", "y", and optionally "col", to give a HEX colour to each region.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a new `ReMIXTURE`` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run"></a>}}
\if{latex}{\out{\hypertarget{method-run}{}}}
\subsection{Method \code{run()}}{
Run the ReMIXTURE analysis. Requires the information table to have been provided upon initialisation or later with $info_table().
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReMIXTURE$run(iterations = 1000, resample = c(0, 0), parallelize = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{iterations}}{The number of samplings requested.}

\item{\code{resample}}{A length two numeric vector in the format c(`iterations`,`proportion`). If `iterations`>100 and 0.1<=`proportion`<=0.9, then $run_resampling() will be automatically run with `iterations` and `proportions` as arguments.}

\item{\code{parallelize}}{An integer indicating how many cores are available to use. Parallel processing is available on Linux-based systems only currently. Leave as default (1L) to disable parallisation.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Nothing
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run_resampling"></a>}}
\if{latex}{\out{\hypertarget{method-run_resampling}{}}}
\subsection{Method \code{run_resampling()}}{
Run the resampling of ReMIXTURE analysis results. Requires the information table to have been provided upon initialisation or later with $info_table().
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReMIXTURE$run_resampling(iterations = 1000, qualifier = 0.9)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{iterations}}{The number of samplings requested.}

\item{\code{qualifier}}{Float number between 0 and 1, which states how many samples to retrieve from every iteration.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A sense of profound satisfaction.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_heatmap"></a>}}
\if{latex}{\out{\hypertarget{method-plot_heatmap}{}}}
\subsection{Method \code{plot_heatmap()}}{
Plot the heatmap of RGOs. Requires that $run() has been called.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReMIXTURE$plot_heatmap()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A ggplot2 plot object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_maps"></a>}}
\if{latex}{\out{\hypertarget{method-plot_maps}{}}}
\subsection{Method \code{plot_maps()}}{
Plot all inter-region RGOs on a map. Requires that $run() has been called.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReMIXTURE$plot_maps(lat_angle = 30, lon_angle = 40)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lat_angle}}{A value indicating how the globe should be rotated along the latitude axis, in standard latitude values in decimal format as you could read off Google maps}

\item{\code{lon_angle}}{A value indicating how the globe should be rotated along the longitude axis, in standard longitude values in decimal format as you could read off Google maps}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReMIXTURE$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
