% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReMIXTURE.R, R/class.R
\docType{package}
\name{ReMIXTURE}
\alias{ReMIXTURE}
\title{ReMIXTURE: Regionwise similarity analysis using a resampling/heirarchical clustering based method.}
\section{Warning}{

This is under development, in early stages.
This version is significantly updated and differs non-trivially from the version presented in the paper Tripodi & Rabanus-Wallace et. al. 2021 about peppers. If you wish to use this method, please contact me to talk about how to cite it. A publication is coming soon.
}

\examples{

my_analysis <- ReMIXTURE$new(
distance_matrix = ReMIXTURE::ReMIXTURE_example_distance_matrix,
region_table = ReMIXTURE::ReMIXTURE_example_region_positions
)

my_analysis$run(iterations = 500,subsample_proportion = 0.8,h_cutoff = 0.013)

my_analysis$plot_heatmap()

my_analysis$plot_maps(alpha_norm_per_region = T)

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-run}{\code{ReMIXTURE$run()}}
\item \href{#method-new}{\code{ReMIXTURE$new()}}
\item \href{#method-plot_heatmap}{\code{ReMIXTURE$plot_heatmap()}}
\item \href{#method-plot_maps}{\code{ReMIXTURE$plot_maps()}}
\item \href{#method-clone}{\code{ReMIXTURE$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run"></a>}}
\if{latex}{\out{\hypertarget{method-run}{}}}
\subsection{Method \code{run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReMIXTURE$run(
  iterations = 2500,
  subsample_proportion = 0.8,
  h_cutoff = private$hcut
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new ReMIXTURE object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReMIXTURE$new(distance_matrix, region_table)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{distance_matrix}}{An all-vs-all, full numeric distance matrix, with rownames and
colnames giving the region of origin of the corresponding individual.}

\item{\code{region_table}}{A data.table describing the lat(y)/long(x)s of each region (numeric or integer values), with columns named "region", "x" and "y".}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a new `ReMIXTURE` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_heatmap"></a>}}
\if{latex}{\out{\hypertarget{method-plot_heatmap}{}}}
\subsection{Method \code{plot_heatmap()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReMIXTURE$plot_heatmap(
  colpalette = colorRampPalette(c("#f2f5ff", "#214feb", "#001261"))(100),
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_maps"></a>}}
\if{latex}{\out{\hypertarget{method-plot_maps}{}}}
\subsection{Method \code{plot_maps()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReMIXTURE$plot_maps(
  alpha_norm_per_region,
  lon_angle = 0,
  lat_angle = 0,
  width_lims = c(5, 35),
  alpha_lims = c(0.05, 0.99)
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReMIXTURE$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
